update: item detail
