+++ src/modules/rules/live/groups/basic.ts
@@ -2,19 +2,6 @@ import { unsafeWindow } from '$'
 import { Item } from '@/types/item'
 import { waitForHead } from '@/utils/init'
 import { logger } from '@/utils/logger'
-import { isEditableElement } from '@/utils/tool'
-
-// 拦截 G 键关注
-const handleKeyGPress = (e: KeyboardEvent) => {
-    if (e.key.toLocaleLowerCase() !== 'g' || e.ctrlKey || e.altKey || e.metaKey) {
-        return
-    }
-    if (isEditableElement(e.target as Element)) {
-        return
-    }
-    e.stopImmediatePropagation()
-    e.preventDefault()
-}
 
 let observer: MutationObserver | undefined
 
@@ -60,21 +47,7 @@ export const liveBasicItems: Item[] = [
         id: 'live-page-remove-wallpaper',
         name: '禁用 直播背景',
     },
-    {
-        type: 'switch',
-        id: 'live-page-disable-hotkey-g-follow',
-        name: '禁用 快捷键 G 关注主播',
-        enableFn: () => {
-            for (const type of ['keydown', 'keypress', 'keyup'] as const) {
-                window.addEventListener(type, handleKeyGPress, true)
-            }
-        },
-        disableFn: () => {
-            for (const type of ['keydown', 'keypress', 'keyup'] as const) {
-                window.removeEventListener(type, handleKeyGPress, true)
-            }
-        },
-    },
+
     {
         type: 'switch',
         id: 'activity-live-auto-jump',
