+++ src/modules/rules/live/groups/info.scss
@@ -5,7 +5,7 @@ html[live-page-head-info-avatar-pendant] {
     }
 }
 
-// 隐藏 粉丝团
+// 隐藏 关注主播/加粉丝团
 html[live-page-head-info-vm-upper-row-follow-ctnr] {
     #head-info-vm .left-anchor-section .follow-ctnr {
         display: none !important;
